improve stest
