--- dapp-chain/side-chain/src/test/java/stest/tron/wallet/common/deposit/MappingTrc20001.java
@@ -1,264 +1,0 @@
-package stest.tron.wallet.common.deposit;
-
-import com.google.protobuf.ByteString;
-import io.grpc.ManagedChannel;
-import io.grpc.ManagedChannelBuilder;
-import java.util.Optional;
-import java.util.concurrent.TimeUnit;
-import lombok.extern.slf4j.Slf4j;
-import org.junit.Assert;
-import org.spongycastle.util.encoders.Hex;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.BeforeSuite;
-import org.testng.annotations.Test;
-import org.tron.api.WalletGrpc;
-import org.tron.api.WalletSolidityGrpc;
-import org.tron.common.crypto.ECKey;
-import org.tron.common.utils.ByteArray;
-import org.tron.common.utils.Utils;
-import org.tron.core.Wallet;
-import org.tron.protos.Protocol.Account;
-import org.tron.protos.Protocol.TransactionInfo;
-import stest.tron.wallet.common.client.Configuration;
-import stest.tron.wallet.common.client.Parameter.CommonConstant;
-import stest.tron.wallet.common.client.WalletClient;
-import stest.tron.wallet.common.client.utils.AbiUtil;
-import stest.tron.wallet.common.client.utils.Base58;
-import stest.tron.wallet.common.client.utils.PublicMethed;
-
-@Slf4j
-public class MappingTrc20001 {
-
-
-  private final String testDepositTrx = Configuration.getByPath("testng.conf")
-      .getString("foundationAccount.key2");
-  private final byte[] testDepositAddress = PublicMethed.getFinalAddress(testDepositTrx);
-  private Long maxFeeLimit = Configuration.getByPath("testng.conf")
-      .getLong("defaultParameter.maxFeeLimit");
-  private ManagedChannel channelSolidity = null;
-  private static final long now = System.currentTimeMillis();
-  private static String tokenName = "testAssetIssue_" + Long.toString(now);
-  private static ByteString assetAccountId = null;
-  private static final long TotalSupply = 1000L;
-  private String description = Configuration.getByPath("testng.conf")
-      .getString("defaultParameter.assetDescription");
-  private String url = Configuration.getByPath("testng.conf")
-      .getString("defaultParameter.assetUrl");
-  private ManagedChannel channelFull = null;
-  private WalletGrpc.WalletBlockingStub blockingStubFull = null;
-
-  private ManagedChannel channelFull1 = null;
-  private WalletGrpc.WalletBlockingStub blockingSideStubFull = null;
-
-
-  private WalletSolidityGrpc.WalletSolidityBlockingStub blockingStubSolidity = null;
-
-  private String fullnode = Configuration.getByPath("testng.conf")
-      .getStringList("fullnode.ip.list").get(0);
-  private String fullnode1 = Configuration.getByPath("testng.conf")
-      .getStringList("fullnode.ip.list").get(1);
-
-  ECKey ecKey1 = new ECKey(Utils.getRandom());
-  byte[] depositAddress = ecKey1.getAddress();
-  String testKeyFordeposit = ByteArray.toHexString(ecKey1.getPrivKeyBytes());
-
-  String mainChainAddress = Configuration.getByPath("testng.conf")
-      .getString("gateway_address.key1");
-  final byte[] mainChainAddressKey = WalletClient.decodeFromBase58Check(mainChainAddress);
-
-  String sideChainAddress = Configuration.getByPath("testng.conf")
-      .getString("gateway_address.key2");
-  final byte[] sideChainAddressKey = WalletClient.decodeFromBase58Check(sideChainAddress);
-
-  @BeforeSuite
-  public void beforeSuite() {
-    Wallet wallet = new Wallet();
-    Wallet.setAddressPreFixByte(CommonConstant.ADD_PRE_FIX_BYTE_MAINNET);
-  }
-
-  /**
-   * constructor.
-   */
-
-  @BeforeClass(enabled = true)
-  public void beforeClass() {
-//    PublicMethed.printAddress(testKeyFordeposit);
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-        .usePlaintext(true)
-        .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-    channelFull1 = ManagedChannelBuilder.forTarget(fullnode1)
-        .usePlaintext(true)
-        .build();
-    blockingSideStubFull = WalletGrpc.newBlockingStub(channelFull1);
-  }
-
-  @Test(enabled = true, description = "Deposit Trc20")
-  public void test1MappingTrc20001() {
-
-    PublicMethed.printAddress(testKeyFordeposit);
-
-    Assert.assertTrue(PublicMethed
-        .sendcoin(depositAddress, 11000_000_000L, testDepositAddress, testDepositTrx,
-            blockingStubFull));
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-
-    String methodStr = "depositTRX()";
-    byte[] input = Hex.decode(AbiUtil.parseMethod(methodStr, "", false));
-
-    Account accountAfter = PublicMethed.queryAccount(depositAddress, blockingStubFull);
-    long accountAfterBalance = accountAfter.getBalance();
-    logger.info("accountAfterBalance:" + accountAfterBalance);
-    Account accountSideAfter = PublicMethed.queryAccount(depositAddress, blockingSideStubFull);
-    long accountSideAfterBalance = accountSideAfter.getBalance();
-    logger.info("accountSideAfterBalance:" + accountSideAfterBalance);
-
-    long callValue = 1000_000_000L;
-    String txid = PublicMethed.triggerContract(mainChainAddressKey, callValue, input,
-        maxFeeLimit, 0, "", depositAddress, testKeyFordeposit, blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingSideStubFull);
-
-    Optional<TransactionInfo> infoById = PublicMethed
-        .getTransactionInfoById(txid, blockingStubFull);
-    Assert.assertEquals(0, infoById.get().getResultValue());
-    long fee = infoById.get().getFee();
-
-    Account accountBefore = PublicMethed.queryAccount(depositAddress, blockingStubFull);
-    long accountBeforeBalance = accountBefore.getBalance();
-    Account accountSideBefore = PublicMethed.queryAccount(depositAddress, blockingSideStubFull);
-    long accountSideBeforeBalance = accountSideBefore.getBalance();
-
-    Assert.assertEquals(0, infoById.get().getResultValue());
-    Assert.assertEquals(10000_000_000L - fee, accountBeforeBalance);
-    Assert.assertEquals(callValue, accountSideBeforeBalance);
-
-    String contractName = "trc20Contract";
-    String code = Configuration.getByPath("testng.conf")
-        .getString("code.code_ContractTRC20");
-    String abi = Configuration.getByPath("testng.conf")
-        .getString("abi.abi_ContractTRC20");
-    String parame = "\"" + Base58.encode58Check(depositAddress) + "\"";
-
-    String deployTxid = PublicMethed
-        .deployContractWithConstantParame(contractName, abi, code, "TronToken(address)",
-            parame, "",
-            maxFeeLimit,
-            0L, 100, null, testKeyFordeposit, depositAddress
-            , blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-
-    infoById = PublicMethed
-        .getTransactionInfoById(deployTxid, blockingStubFull);
-    byte[] trc20Contract = infoById.get().getContractAddress().toByteArray();
-    Assert.assertEquals(0, infoById.get().getResultValue());
-    Assert.assertNotNull(trc20Contract);
-
-    //account is other
-
-    String mapTxid1 = PublicMethed
-        .mappingTrc20(mainChainAddressKey, deployTxid, 1000000000,
-            testDepositAddress, testDepositTrx, blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingSideStubFull);
-
-    Optional<TransactionInfo> infoById1 = PublicMethed
-        .getTransactionInfoById(mapTxid1, blockingStubFull);
-    Assert.assertEquals("FAILED", infoById1.get().getResult().name());
-    Assert.assertEquals(1, infoById1.get().getResultValue());
-
-    //account is not exist
-
-    ECKey ecKey2 = new ECKey(Utils.getRandom());
-    byte[] depositAddress2 = ecKey2.getAddress();
-    String testKeyFordeposit2 = ByteArray.toHexString(ecKey2.getPrivKeyBytes());
-
-    String s = PublicMethed
-        .mappingTrc20(mainChainAddressKey, deployTxid, 1000000000,
-            depositAddress2, testKeyFordeposit2, blockingStubFull);
-
-//not exist txid
-    String deployTxidFake = "1111111";
-
-    String mapTxid2 = PublicMethed
-        .mappingTrc20(mainChainAddressKey, deployTxidFake, 1000000000,
-            testDepositAddress, testDepositTrx, blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingSideStubFull);
-
-    Optional<TransactionInfo> infoById2 = PublicMethed
-        .getTransactionInfoById(mapTxid2, blockingStubFull);
-    Assert.assertEquals("FAILED", infoById2.get().getResult().name());
-    Assert.assertEquals(1, infoById2.get().getResultValue());
-    Assert.assertNotNull(mapTxid2);
-
-    //other txid
-    String mapTxid3 = PublicMethed
-        .mappingTrc20(mainChainAddressKey, txid, 1000000000,
-            testDepositAddress, testDepositTrx, blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingSideStubFull);
-
-    Optional<TransactionInfo> infoById3 = PublicMethed
-        .getTransactionInfoById(mapTxid3, blockingStubFull);
-    Assert.assertEquals("FAILED", infoById3.get().getResult().name());
-    Assert.assertEquals(1, infoById3.get().getResultValue());
-    Assert.assertNotNull(mapTxid3);
-
-    //feelimit is 1
-    String mapTxid4 = PublicMethed
-        .mappingTrc20(mainChainAddressKey, deployTxid, 1,
-            depositAddress, testKeyFordeposit, blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingSideStubFull);
-
-    Optional<TransactionInfo> infoById4 = PublicMethed
-        .getTransactionInfoById(mapTxid4, blockingStubFull);
-    Assert.assertEquals("FAILED", infoById3.get().getResult().name());
-    Assert.assertEquals(1, infoById3.get().getResultValue());
-    Assert.assertNotNull(mapTxid4);
-
-    //success ,again map
-
-    String mapTxid5 = PublicMethed
-        .mappingTrc20(mainChainAddressKey, deployTxid, 1000000000,
-            depositAddress, testKeyFordeposit, blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingSideStubFull);
-
-    Optional<TransactionInfo> infoById5 = PublicMethed
-        .getTransactionInfoById(mapTxid5, blockingStubFull);
-    Assert.assertEquals("SUCESS", infoById5.get().getResult().name());
-    Assert.assertEquals(0, infoById5.get().getResultValue());
-
-    String mapTxid6 = PublicMethed
-        .mappingTrc20(mainChainAddressKey, deployTxid, 1000000000,
-            depositAddress, testKeyFordeposit, blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingSideStubFull);
-
-    Optional<TransactionInfo> infoById6 = PublicMethed
-        .getTransactionInfoById(mapTxid6, blockingStubFull);
-    Assert.assertEquals("FAILED", infoById6.get().getResult().name());
-    Assert.assertEquals(1, infoById6.get().getResultValue());
-    Assert.assertNotNull(mapTxid6);
-
-
-  }
-
-
-  /**
-   * constructor.
-   */
-  @AfterClass
-  public void shutdown() throws InterruptedException {
-    if (channelFull != null) {
-      channelFull.shutdown().awaitTermination(5, TimeUnit.SECONDS);
-    }
-    if (channelFull1 != null) {
-      channelFull1.shutdown().awaitTermination(5, TimeUnit.SECONDS);
-    }
-  }
-
-}
