+++ dapp-chain/side-chain/src/test/java/stest/tron/wallet/depositWithdraw/MappingTrc20001.java
@@ -1,6 +1,5 @@
 package stest.tron.wallet.common.deposit;
 
-import com.google.protobuf.ByteString;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import java.util.Optional;
@@ -12,7 +11,6 @@ import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
-import org.tron.api.GrpcAPI.TransactionExtention;
 import org.tron.api.WalletGrpc;
 import org.tron.api.WalletSolidityGrpc;
 import org.tron.common.crypto.ECKey;
@@ -29,7 +27,7 @@ import stest.tron.wallet.common.client.utils.Base58;
 import stest.tron.wallet.common.client.utils.PublicMethed;
 
 @Slf4j
-public class DepositTrc721001 {
+public class MappingTrc20001 {
 
 
   private final String testDepositTrx = Configuration.getByPath("testng.conf")
@@ -38,14 +36,6 @@ public class DepositTrc721001 {
   private Long maxFeeLimit = Configuration.getByPath("testng.conf")
       .getLong("defaultParameter.maxFeeLimit");
   private ManagedChannel channelSolidity = null;
-  private static final long now = System.currentTimeMillis();
-  private static String tokenName = "testAssetIssue_" + Long.toString(now);
-  private static ByteString assetAccountId = null;
-  private static final long TotalSupply = 1000L;
-  private String description = Configuration.getByPath("testng.conf")
-      .getString("defaultParameter.assetDescription");
-  private String url = Configuration.getByPath("testng.conf")
-      .getString("defaultParameter.assetUrl");
   private ManagedChannel channelFull = null;
   private WalletGrpc.WalletBlockingStub blockingStubFull = null;
 
@@ -56,9 +46,9 @@ public class DepositTrc721001 {
   private WalletSolidityGrpc.WalletSolidityBlockingStub blockingStubSolidity = null;
 
   private String fullnode = Configuration.getByPath("testng.conf")
-      .getStringList("fullnode.ip.list").get(0);
+      .getStringList("mainfullnode.ip.list").get(0);
   private String fullnode1 = Configuration.getByPath("testng.conf")
-      .getStringList("fullnode.ip.list").get(1);
+      .getStringList("fullnode.ip.list").get(0);
 
   ECKey ecKey1 = new ECKey(Utils.getRandom());
   byte[] depositAddress = ecKey1.getAddress();
@@ -84,7 +74,7 @@ public class DepositTrc721001 {
 
   @BeforeClass(enabled = true)
   public void beforeClass() {
-//    PublicMethed.printAddress(depositKey001);
+//    PublicMethed.printAddress(testKeyFordeposit);
     channelFull = ManagedChannelBuilder.forTarget(fullnode)
         .usePlaintext(true)
         .build();
@@ -95,8 +85,8 @@ public class DepositTrc721001 {
     blockingSideStubFull = WalletGrpc.newBlockingStub(channelFull1);
   }
 
-  @Test(enabled = true, description = "Deposit Trc721")
-  public void test1DepositTrc20001() {
+  @Test(enabled = true, description = "Deposit Trc20")
+  public void test1MappingTrc20001() {
 
     PublicMethed.printAddress(testKeyFordeposit);
 
@@ -135,16 +125,15 @@ public class DepositTrc721001 {
     Assert.assertEquals(10000_000_000L - fee, accountBeforeBalance);
     Assert.assertEquals(callValue, accountSideBeforeBalance);
 
-    String contractName = "trc721";
+    String contractName = "trc20Contract";
     String code = Configuration.getByPath("testng.conf")
-        .getString("code.code_ContractTRC721");
+        .getString("code.code_ContractTRC20");
     String abi = Configuration.getByPath("testng.conf")
-        .getString("abi.abi_ContractTRC721");
-    String parame = "\"" + Base58.encode58Check(depositAddress) + "\",\"nmb721wm\",\"nmbwm\"";
+        .getString("abi.abi_ContractTRC20");
+    String parame = "\"" + Base58.encode58Check(depositAddress) + "\"";
 
     String deployTxid = PublicMethed
-        .deployContractWithConstantParame(contractName, abi, code,
-            "constructor(address,string,string)",
+        .deployContractWithConstantParame(contractName, abi, code, "TronToken(address)",
             parame, "",
             maxFeeLimit,
             0L, 100, null, testKeyFordeposit, depositAddress
@@ -157,80 +146,96 @@ public class DepositTrc721001 {
     Assert.assertEquals(0, infoById.get().getResultValue());
     Assert.assertNotNull(trc20Contract);
 
-    String parame1 = "\"" + Base58.encode58Check(depositAddress) + "\"," + 1001;
-    String mintTxid = PublicMethed
-        .triggerContract(trc20Contract, "mint(address,uint256)", parame1, false, 0, maxFeeLimit,
-            depositAddress, testKeyFordeposit, blockingStubFull);
-    infoById = PublicMethed.getTransactionInfoById(mintTxid, blockingStubFull);
-    Assert.assertNotNull(mintTxid);
-    Assert.assertEquals(0, infoById.get().getResultValue());
-    Assert.assertEquals("SUCESS", infoById.get().getResult().name());
+    //account is other
 
-    String mapTxid = PublicMethed
-        .mappingTrc721(mainChainAddressKey, deployTxid, 1000000000,
-            depositAddress, testKeyFordeposit, blockingStubFull);
+    String mapTxid1 = PublicMethed
+        .mappingTrc20(mainChainAddressKey, deployTxid, 1000000000,
+            testDepositAddress, testDepositTrx, blockingStubFull);
     PublicMethed.waitProduceNextBlock(blockingStubFull);
     PublicMethed.waitProduceNextBlock(blockingSideStubFull);
 
     Optional<TransactionInfo> infoById1 = PublicMethed
-        .getTransactionInfoById(mapTxid, blockingSideStubFull);
-    Assert.assertEquals("SUCESS", infoById1.get().getResult().name());
-    Assert.assertEquals(0, infoById1.get().getResultValue());
-    Assert.assertNotNull(mapTxid);
+        .getTransactionInfoById(mapTxid1, blockingStubFull);
+    Assert.assertEquals("FAILED", infoById1.get().getResult().name());
+    Assert.assertEquals(1, infoById1.get().getResultValue());
 
-    String parame2 = "\"" + Base58.encode58Check(trc20Contract) + "\"";
-    byte[] input2 = Hex
-        .decode(AbiUtil.parseMethod("mainToSideContractMap(address)", parame2, false));
-    TransactionExtention return1 = PublicMethed
-        .triggerContractForTransactionExtention(sideChainAddressKey, 0, input2,
-            maxFeeLimit,
-            0, "0",
-            depositAddress, testKeyFordeposit, blockingSideStubFull);
-    infoById = PublicMethed
-        .getTransactionInfoById(txid, blockingSideStubFull);
-    logger.info(Hex.toHexString(return1.getConstantResult(0).toByteArray()));
-    String ContractRestule = Hex.toHexString(return1.getConstantResult(0).toByteArray());
+    //account is not exist
 
-    String tmpAddress = ContractRestule.substring(24);
-    logger.info(tmpAddress);
-    String addressHex = "41" + tmpAddress;
-    logger.info("address_hex: " + addressHex);
-    String addressFinal = Base58.encode58Check(ByteArray.fromHexString(addressHex));
-    logger.info("address_final: " + addressFinal);
+    ECKey ecKey2 = new ECKey(Utils.getRandom());
+    byte[] depositAddress2 = ecKey2.getAddress();
+    String testKeyFordeposit2 = ByteArray.toHexString(ecKey2.getPrivKeyBytes());
 
-    byte[] sideContractAddress = WalletClient.decodeFromBase58Check(addressFinal);
-    Assert.assertEquals(0, infoById.get().getResultValue());
-    Assert.assertNotNull(sideContractAddress);
+    String s = PublicMethed
+        .mappingTrc20(mainChainAddressKey, deployTxid, 1000000000,
+            depositAddress2, testKeyFordeposit2, blockingStubFull);
 
-    String deposittrx = PublicMethed
-        .depositTrc721(WalletClient.encode58Check(trc20Contract), mainChainAddress, 1001,
-            1000000000,
-            depositAddress, testKeyFordeposit, blockingStubFull);
-    logger.info(deposittrx);
+//not exist txid
+    String deployTxidFake = "1111111";
+
+    String mapTxid2 = PublicMethed
+        .mappingTrc20(mainChainAddressKey, deployTxidFake, 1000000000,
+            testDepositAddress, testDepositTrx, blockingStubFull);
     PublicMethed.waitProduceNextBlock(blockingStubFull);
     PublicMethed.waitProduceNextBlock(blockingSideStubFull);
-    infoById = PublicMethed.getTransactionInfoById(deposittrx, blockingStubFull);
 
-    String arg = "1001";
-    byte[] input1 = Hex.decode(AbiUtil.parseMethod("ownerOf(uint256)", arg, false));
-    String ownerTrx = PublicMethed
-        .triggerContractSideChain(sideContractAddress, mainChainAddressKey, 0l, input1, 1000000000,
-            0l, "0", depositAddress, testKeyFordeposit, blockingSideStubFull);
-    logger.info("ownerTrx : " + ownerTrx);
     Optional<TransactionInfo> infoById2 = PublicMethed
-        .getTransactionInfoById(ownerTrx, blockingSideStubFull);
-    tmpAddress = ByteArray.toHexString(infoById2.get().getContractResult(0).toByteArray());
-    tmpAddress = tmpAddress.substring(24);
-    addressHex = "41" + tmpAddress;
-    logger.info("address_hex: " + addressHex);
-    addressFinal = Base58.encode58Check(ByteArray.fromHexString(addressHex));
-    logger.info("address_final: " + addressFinal);
+        .getTransactionInfoById(mapTxid2, blockingStubFull);
+    Assert.assertEquals("FAILED", infoById2.get().getResult().name());
+    Assert.assertEquals(1, infoById2.get().getResultValue());
+    Assert.assertNotNull(mapTxid2);
 
-    Assert.assertEquals("SUCESS", infoById2.get().getResult().name());
-    Assert.assertEquals(0, infoById2.get().getResultValue());
-    Assert.assertNotNull(ownerTrx);
+    //other txid
+    String mapTxid3 = PublicMethed
+        .mappingTrc20(mainChainAddressKey, txid, 1000000000,
+            testDepositAddress, testDepositTrx, blockingStubFull);
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+    PublicMethed.waitProduceNextBlock(blockingSideStubFull);
 
-    Assert.assertEquals(Base58.encode58Check(depositAddress), addressFinal);
+    Optional<TransactionInfo> infoById3 = PublicMethed
+        .getTransactionInfoById(mapTxid3, blockingStubFull);
+    Assert.assertEquals("FAILED", infoById3.get().getResult().name());
+    Assert.assertEquals(1, infoById3.get().getResultValue());
+    Assert.assertNotNull(mapTxid3);
+
+    //feelimit is 1
+    String mapTxid4 = PublicMethed
+        .mappingTrc20(mainChainAddressKey, deployTxid, 1,
+            depositAddress, testKeyFordeposit, blockingStubFull);
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+    PublicMethed.waitProduceNextBlock(blockingSideStubFull);
+
+    Optional<TransactionInfo> infoById4 = PublicMethed
+        .getTransactionInfoById(mapTxid4, blockingStubFull);
+    Assert.assertEquals("FAILED", infoById3.get().getResult().name());
+    Assert.assertEquals(1, infoById3.get().getResultValue());
+    Assert.assertNotNull(mapTxid4);
+
+    //success ,again map
+
+    String mapTxid5 = PublicMethed
+        .mappingTrc20(mainChainAddressKey, deployTxid, 1000000000,
+            depositAddress, testKeyFordeposit, blockingStubFull);
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+    PublicMethed.waitProduceNextBlock(blockingSideStubFull);
+
+    Optional<TransactionInfo> infoById5 = PublicMethed
+        .getTransactionInfoById(mapTxid5, blockingStubFull);
+    Assert.assertEquals("SUCESS", infoById5.get().getResult().name());
+    Assert.assertEquals(0, infoById5.get().getResultValue());
+
+    String mapTxid6 = PublicMethed
+        .mappingTrc20(mainChainAddressKey, deployTxid, 1000000000,
+            depositAddress, testKeyFordeposit, blockingStubFull);
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+    PublicMethed.waitProduceNextBlock(blockingSideStubFull);
+
+    Optional<TransactionInfo> infoById6 = PublicMethed
+        .getTransactionInfoById(mapTxid6, blockingStubFull);
+    Assert.assertEquals("FAILED", infoById6.get().getResult().name());
+    Assert.assertEquals(1, infoById6.get().getResultValue());
+    Assert.assertNotNull(mapTxid6);
+
+
   }
